Organize tool marketing landing views
